Move handlers and project functions into ops module
