+++ src/bin/mgmt-configs/main.rs
@@ -1,12 +1,12 @@
 use std::path::PathBuf;
 
 use mgmt::config_values::config;
-use mgmt::db::{self, Configuration, LoadFromConfiguration};
+use mgmt::db;
+use mgmt::ops;
 
 use anyhow::anyhow;
 use clap::{arg, ArgAction, Command};
 use sqlx::mysql::MySqlPoolOptions;
-use sqlx::{MySql, Pool};
 
 /**
  * Sets up the CLI for the mgmt-configs binary.
@@ -300,378 +300,6 @@ fn cli() -> Command {
         )
 }
 
-/**
- * Handler for the `mgmt-configs env populate` command.
- */
-async fn populate_env(pool: &Pool<MySql>) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let mut env_config = config::ConfigValues::default();
-    env_config.ask_for_info(&mut tx).await?;
-    tx.commit().await?;
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs sections add` command.
- */
-async fn add_section(pool: &Pool<MySql>, section: &str) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    db::add_section(&mut tx, &section).await?;
-    tx.commit().await?;
-    Ok(())
-}
-/**
- * Handler for the `mgmt-configs sections delete` command.
- */
-async fn delete_section(pool: &Pool<MySql>, section: &str) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    db::delete_section(&mut tx, &section).await?;
-    tx.commit().await?;
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs sections list` command.
- */
-async fn list_sections(pool: &Pool<MySql>) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let sections = db::list_sections(&mut tx).await?;
-    tx.commit().await?;
-    for section in sections {
-        println!("{}", section);
-    }
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs defaults set` command.
- */
-async fn set_default_value(
-    pool: &Pool<MySql>,
-    section: &str,
-    key: &str,
-    value: &str,
-    value_type: &str,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let has_section = db::has_section(&mut tx, section).await?;
-    if has_section {
-        let cfg_id =
-            db::set_default_config_value(&mut tx, section, &key, &value, &value_type).await?;
-        println!("Added default config value with and ID of {}", cfg_id);
-    } else {
-        return Err(anyhow!("No section found with name: {section}"));
-    }
-    tx.commit().await?;
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs defaults get` command.
- */
-async fn get_default_value(pool: &Pool<MySql>, section: &str, key: &str) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let cfg: db::Configuration;
-    let has_section = db::has_section(&mut tx, section).await?;
-    if has_section {
-        let has_default_value = db::has_default_config_value(&mut tx, section, key).await?;
-        if has_default_value {
-            cfg = db::get_default_config_value(&mut tx, section, key).await?;
-        } else {
-            return Err(anyhow!(
-                "No default value found for section: {section}, key: {key}"
-            ));
-        }
-    } else {
-        return Err(anyhow!("No section found with name: {section}"));
-    }
-    tx.commit().await?;
-    println!("{:?}", cfg);
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs defaults delete` command.
- */
-async fn delete_default_value(pool: &Pool<MySql>, section: &str, key: &str) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let has_section = db::has_section(&mut tx, section).await?;
-    if has_section {
-        let has_default_value = db::has_default_config_value(&mut tx, section, key).await?;
-        if has_default_value {
-            db::delete_default_config_value(&mut tx, section, key).await?;
-        } else {
-            return Err(anyhow!(
-                "No default value found for section: {section}, key: {key}"
-            ));
-        }
-    } else {
-        return Err(anyhow!("No section found with name: {section}"));
-    }
-    tx.commit().await?;
-    println!("Deleted default value: {}.{}", section, key);
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs defaults list` command.
- */
-async fn list_default_values(
-    pool: &Pool<MySql>,
-    section: Option<&str>,
-    key: Option<&str>,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let cfgs = db::list_default_config_values(&mut tx, section, key).await?;
-    tx.commit().await?;
-    for cfg in cfgs {
-        if let (Some(section), Some(key), Some(value)) = (cfg.section, cfg.key, cfg.value) {
-            println!("{}.{} = {}", section, key, value);
-        }
-    }
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs defaults render` command.
- */
-async fn render_default_values(
-    pool: &Pool<MySql>,
-    output_file: Option<PathBuf>,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-
-    let all_default_cfgs = db::list_default_config_values(&mut tx, None, None).await?;
-    let mut cv = config::ConfigValues::default();
-    let mut section_options = config::SectionOptions::default();
-    section_options.set_all(true)?;
-    cv.set_section_options(section_options);
-    cv.reset_sections()?;
-    cv.cfg_set_keys(all_default_cfgs)?;
-
-    if let Some(output_file) = output_file {
-        let yaml = serde_yaml::to_string(&cv)?;
-        std::fs::write(output_file, yaml)?;
-    } else {
-        let yaml = serde_yaml::to_string(&cv)?;
-        println!("{}", yaml);
-    }
-
-    tx.commit().await?;
-
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs values set` command.
- */
-async fn set_value(
-    pool: &Pool<MySql>,
-    environment: &str,
-    section: &str,
-    key: &str,
-    value: &str,
-    value_type: &str,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-
-    let env_id = db::get_env_id(&mut tx, &environment)
-        .await?
-        .ok_or_else(|| anyhow::anyhow!("No environment found with name: {environment}"))?;
-
-    let has_default = db::has_default_config_value(&mut tx, section, &key).await?;
-    if has_default {
-        let cfg_id = db::set_config_value(&mut tx, section, &key, &value, &value_type).await?;
-        db::add_env_cfg_value(&mut tx, env_id, cfg_id).await?;
-        println!(
-            "Added config value to environment '{}': {}.{} = {}",
-            environment, section, key, value
-        );
-    } else {
-        tx.rollback().await?;
-        return Err(anyhow!(
-            "No default value found for section: {section}, key: {key}"
-        ));
-    }
-
-    tx.commit().await?;
-
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs values get` command.
- */
-async fn get_value(
-    pool: &Pool<MySql>,
-    environment: &str,
-    section: &str,
-    key: &str,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let cfg: db::Configuration;
-
-    let has_config_value = db::has_config_value(&mut tx, environment, section, key).await?;
-    if has_config_value {
-        cfg = db::get_config_value(&mut tx, environment, section, key).await?;
-    } else {
-        let has_default_value = db::has_default_config_value(&mut tx, section, key).await?;
-        if has_default_value {
-            cfg = db::get_default_config_value(&mut tx, section, key).await?;
-        } else {
-            tx.rollback().await?;
-            return Err(anyhow!(
-                "No default value found for section: {section}, key: {key}"
-            ));
-        }
-    }
-    tx.commit().await?;
-    if let (Some(section), Some(key), Some(value)) = (cfg.section, cfg.key, cfg.value) {
-        println!("{}.{} = {}", section, key, value);
-    }
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs values delete` command.
- */
-async fn delete_value(
-    pool: &Pool<MySql>,
-    environment: &str,
-    section: &str,
-    key: &str,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    db::delete_config_value(&mut tx, environment, section, key).await?;
-    tx.commit().await?;
-    println!(
-        "Deleted config value from environment '{}': {}.{}",
-        environment, section, key
-    );
-    Ok(())
-}
-
-/**
- * Handler for the `mgmt-configs values list` command.
- */
-async fn list_values(
-    pool: &Pool<MySql>,
-    environment: Option<&str>,
-    section: Option<&str>,
-    key: Option<&str>,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let cfgs = db::list_config_values(&mut tx, environment, section, key).await?;
-    tx.commit().await?;
-    for cfg in cfgs {
-        if let (Some(section), Some(key), Some(value)) = (cfg.section, cfg.key, cfg.value) {
-            println!("{}.{} = {}", section, key, value);
-        }
-    }
-    Ok(())
-}
-
-/**
- * Handler  for the `mgmt-configs values render` command.
- */
-async fn render_values(
-    pool: &Pool<MySql>,
-    environment: &str,
-    opts: &config::SectionOptions,
-    output_file: Option<PathBuf>,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let mut all_cfgs: Vec<Configuration> = Vec::new();
-
-    let all_default_cfgs = db::list_default_config_values(&mut tx, None, None).await?;
-    for default in all_default_cfgs
-        .into_iter()
-        .filter(|cfg| opts.include_section(&cfg.section.clone().unwrap_or_default()))
-    {
-        if let (Some(section), Some(key)) = (default.section.clone(), default.key.clone()) {
-            let has_config_value = db::has_config_value(&mut tx, environment, &section, &key)
-                .await
-                .unwrap_or(false);
-
-            if has_config_value {
-                all_cfgs.push(
-                    db::get_config_value(&mut tx, environment, &section, &key)
-                        .await
-                        .unwrap(),
-                );
-            } else {
-                all_cfgs.push(default);
-            }
-        }
-    }
-
-    let mut cv = config::ConfigValues::default();
-    cv.set_section_options(opts.clone());
-    cv.reset_sections()?;
-    cv.cfg_set_keys(all_cfgs)?;
-
-    tx.commit().await?;
-
-    if let Some(output_file) = output_file {
-        let yaml = serde_yaml::to_string(&cv)?;
-        std::fs::write(output_file, yaml)?;
-    } else {
-        let yaml = serde_yaml::to_string(&cv)?;
-        println!("{}", yaml);
-    }
-
-    Ok(())
-}
-
-/**
- * Handler for importing files.
- */
-async fn import_yaml_file(
-    pool: &Pool<MySql>,
-    path: PathBuf,
-    environment: &str,
-) -> anyhow::Result<()> {
-    let mut tx = pool.begin().await?;
-    let file = std::fs::File::open(path)?;
-    let cv: config::ConfigValues = serde_yaml::from_reader(file)?;
-    let items: Vec<db::Configuration> = cv.into();
-    for item in items.into_iter() {
-        if let (Some(section), Some(key), Some(value), Some(value_type)) = (
-            item.section.clone(),
-            item.key.clone(),
-            item.value.clone(),
-            item.value_type.clone(),
-        ) {
-            println!("{}.{} = {}", section, key, value);
-            let real_section: String;
-            if section.is_empty() {
-                real_section = "TopLevel".to_string();
-            } else {
-                real_section = section;
-            }
-
-            if db::has_default_config_value(&mut tx, &real_section, &key).await? {
-                let cfg_id =
-                    db::set_config_value(&mut tx, &real_section, &key, &value, &value_type).await?;
-                let env_id = db::get_env_id(&mut tx, &environment)
-                    .await?
-                    .ok_or_else(|| {
-                        anyhow::anyhow!("No environment found with name: {environment}")
-                    })?;
-                db::add_env_cfg_value(&mut tx, env_id, cfg_id).await?;
-            } else {
-                tx.rollback().await?;
-                return Err(anyhow!(
-                    "No default value found for section: {real_section}, key: {key}"
-                ));
-            }
-        }
-    }
-
-    tx.commit().await?;
-    Ok(())
-}
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     let command = cli().get_matches();
@@ -695,7 +323,7 @@ async fn main() -> anyhow::Result<()> {
 
             match create_cmd {
                 ("populate", _) => {
-                    populate_env(&pool).await?;
+                    ops::populate_env(&pool).await?;
                 }
 
                 ("create", sub_m) => {
@@ -755,7 +383,7 @@ async fn main() -> anyhow::Result<()> {
                         )
                     })?;
 
-                    add_section(&pool, &section).await?;
+                    ops::add_section(&pool, &section).await?;
                 }
 
                 // Match on the `sections delete` subcommand.
@@ -766,12 +394,12 @@ async fn main() -> anyhow::Result<()> {
                         )
                     })?;
 
-                    delete_section(&pool, &section).await?;
+                    ops::delete_section(&pool, &section).await?;
                 }
 
                 // Match on the `sections list` subcommand.
                 ("list", _) => {
-                    list_sections(&pool).await?;
+                    ops::list_sections(&pool).await?;
                 }
 
                 (name, _) => {
@@ -805,7 +433,7 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No type specified. Use --type <type> to specify a type.")
                     })?;
 
-                    set_default_value(&pool, &section, &key, &value, &value_type).await?;
+                    ops::set_default_value(&pool, &section, &key, &value, &value_type).await?;
                 }
 
                 ("get", sub_m) => {
@@ -819,7 +447,7 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No key specified. Use --key <key> to specify a key.")
                     })?;
 
-                    get_default_value(&pool, &section, &key).await?;
+                    ops::get_default_value(&pool, &section, &key).await?;
                 }
 
                 ("delete", sub_m) => {
@@ -833,7 +461,7 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No key specified. Use --key <key> to specify a key.")
                     })?;
 
-                    delete_default_value(&pool, &section, &key).await?;
+                    ops::delete_default_value(&pool, &section, &key).await?;
                 }
 
                 ("list", sub_m) => {
@@ -847,12 +475,12 @@ async fn main() -> anyhow::Result<()> {
                         None => None,
                     };
 
-                    list_default_values(&pool, section, key).await?;
+                    ops::list_default_values(&pool, section, key).await?;
                 }
 
                 ("render", sub_m) => {
                     let output_file = sub_m.get_one::<PathBuf>("file").cloned();
-                    render_default_values(&pool, output_file).await?;
+                    ops::render_default_values(&pool, output_file).await?;
                 }
 
                 (name, _) => {
@@ -893,7 +521,8 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No type specified. Use --type <type> to specify a type.")
                     })?;
 
-                    set_value(&pool, &environment, &section, &key, &value, &value_type).await?;
+                    ops::set_value(&pool, &environment, &section, &key, &value, &value_type)
+                        .await?;
                 }
 
                 // Match on the `values get` subcommand.
@@ -914,7 +543,7 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No key specified. Use --key <key> to specify a key.")
                     })?;
 
-                    get_value(&pool, &environment, &section, &key).await?;
+                    ops::get_value(&pool, &environment, &section, &key).await?;
                 }
 
                 // Match on the `values delete` subcommand.
@@ -935,7 +564,7 @@ async fn main() -> anyhow::Result<()> {
                         anyhow!("No key specified. Use --key <key> to specify a key.")
                     })?;
 
-                    delete_value(&pool, &environment, &section, &key).await?;
+                    ops::delete_value(&pool, &environment, &section, &key).await?;
                 }
 
                 // Match on the `values list` subcommand.
@@ -955,7 +584,7 @@ async fn main() -> anyhow::Result<()> {
                         None => None,
                     };
 
-                    list_values(&pool, environment, section, key).await?;
+                    ops::list_values(&pool, environment, section, key).await?;
                 }
 
                 ("render", sub_m) => {
@@ -969,7 +598,7 @@ async fn main() -> anyhow::Result<()> {
                         None => None,
                     };
                     let opts = config::SectionOptions::new(sub_m);
-                    render_values(&pool, &environment, &opts, output_file).await?;
+                    ops::render_values(&pool, &environment, &opts, output_file).await?;
                 }
 
                 ("import", sub_m) => {
@@ -983,7 +612,7 @@ async fn main() -> anyhow::Result<()> {
                         )
                     })?;
 
-                    import_yaml_file(&pool, path.to_path_buf(), environment).await?;
+                    ops::import_yaml_file(&pool, path.to_path_buf(), environment).await?;
                 }
 
                 (name, _) => {
