+++ src/lib.rs
@@ -5,4 +5,4 @@ pub mod configs;
 pub mod db;
 pub mod dolt;
 pub mod git;
-pub mod project;
+pub mod ops;
